vanishingPoints.py
Vanishing point detection by Hough Transform. At the moment, it doesn't work properly there is something to be done as reported in the code in order to improve the solution. Moreover, it may be required a change of approach to take multiple vanishing points into account, since at this stage the detection is meant to focus on one single vanishing point